lib/ukboot: Declare `main` as a strong symbol rather than a weak one
When a symbol is _declared_ weak, `ld` (at least GNU `ld`) will set it
to `NULL` if the symbol cannot be found in the `.o` files (even if it
could find it in a `.a` library). As `do_main` calls `main`
inconditionnally (ie without testing whether it is available or not),
this commit makes the declaration of the symbol strong, so that the
linker fails if no `main` is found.

Signed-off-by: Samuel Hym <samuel@tarides.com>