drivers/xencons: Rename `libxenconsole` to `libxencons`
Rename the driver library and driver directory to better match
the naming of Xen. The name "xencons" is derived from the naming
of the device by the official Xen headers. Configuration options
are adopted accordingly.

In order to reduce potential name clashes, the exported symbol
`flush_console()` is renamed to `xencons_flush()`.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1501