build, plat/*: Ensure building only a single platform
As part of the re-arch efforts and driver re-organization, the
re-arch team decided to no longer support creating multiple target
images with a single configuration and single `make` call. This is
to simplify the APIs, allow further specialization, and even allow
calling a driver API directly from application code.
As a temporary workaround, we stop a build when more than one
platform is selected and display a warning in the platform
configuration menu. This ensures backwards compatibility during
re-arch transition. In the longer term, we plan to restructure
the platform configuration submenu with a `choice` or similar.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1501