drivers/ukconsole/vga: Add initial sanity check
On initialization, ensure that the VGA controller is present and in
the expected state, otherwise cancel the initialization and
registration of the VGA driver, but continue booting.

For platforms without an VGA controller (such as firecracker), this
avoids aborts due to an illicit registration of a non-existing VGA
memory region.

Signed-off-by: Florian Rommel <mail@florommel.de>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1530